+++ src/pages/RegisterPage/RegisterPage.tsx
@@ -1,6 +1,6 @@
 import postUser from '@/apis/users/postUser';
 import InputForm from '@/components/common/InputForm/InputForm';
-import { ERROR_MESSAGE } from '@/constants/errorMessage';
+import { ERROR_MESSAGE } from '@/constants/toastMessage';
 import { validateName, validateNumber } from '@/utils/validate';
 
 import { FieldValues, useForm } from 'react-hook-form';
